+++ src/robotics_utils/collision_models/aabb.py
