+++ src/robotics_utils/collision_models/primitive_shapes.py
@@ -7,8 +7,8 @@ from typing import Protocol
 
 import numpy as np
 
-from robotics_utils.kinematics.collision_models.aabb import AxisAlignedBoundingBox
-from robotics_utils.kinematics.point3d import Point3D
+from robotics_utils.collision_models.aabb import AxisAlignedBoundingBox
+from robotics_utils.kinematics import Point3D
 
 
 class PrimitiveShape(Protocol):
