--- src/robotics_utils/filesystem/convert_memory.py
@@ -1,3 +1,0 @@
-"""Define utility functions for converting units of memory."""
-
-Gi_TO_B = 1024**3
