+++ src/robotics_utils/io/yaml_utils.py
@@ -5,12 +5,7 @@ from __future__ import annotations
 from pathlib import Path
 from typing import Any
 
-import trimesh
 import yaml
-
-from robotics_utils.kinematics import DEFAULT_FRAME
-from robotics_utils.kinematics.collision_models.meshes import load_trimesh_from_yaml_data
-from robotics_utils.kinematics.poses import Pose3D
 
 
 def export_yaml_data(data: dict[str, Any] | list[Any], filepath: Path) -> None:
@@ -48,35 +43,3 @@ def load_yaml_data(yaml_path: Path, required_keys: set[str] | None = None) -> An
                 raise KeyError(f"Required key '{key}' was missing in data loaded from {yaml_path}")
 
     return yaml_data
-
-
-def load_named_poses(yaml_path: Path, collection_name: str) -> dict[str, Pose3D]:
-    """Load a collection of named poses from the given YAML file.
-
-    :param yaml_path: Path to a YAML file containing pose data
-    :param collection_name: Name of the collection of poses to be imported (e.g., "object_poses")
-    :return: Dictionary mapping pose-frame names to their imported 3D poses
-    """
-    yaml_data = load_yaml_data(yaml_path, required_keys={collection_name})
-    default_frame = yaml_data.get("default_frame", DEFAULT_FRAME)
-    poses_data: dict[str, Any] = yaml_data[collection_name]
-
-    return {
-        pose_name: Pose3D.from_yaml_data(pose_data, default_frame)
-        for pose_name, pose_data in poses_data.items()
-    }
-
-
-def load_named_mesh(mesh_key: str, yaml_path: Path) -> trimesh.Trimesh:
-    """Load the specified mesh from the given YAML file.
-
-    :param mesh_key: YAML key used to access the imported mesh
-    :param yaml_path: Path to a YAML file specifying mesh data
-    :return: Constructed trimesh.Trimesh instance
-    """
-    yaml_data = load_yaml_data(yaml_path, required_keys={"meshes"})
-    mesh_data = yaml_data["meshes"].get(mesh_key)
-    if mesh_data is None:
-        raise KeyError(f"Could not find mesh named '{mesh_key}' in YAML file {yaml_path}")
-
-    return load_trimesh_from_yaml_data(mesh_data, yaml_path)
