+++ src/robotics_utils/kinematics/__init__.py
@@ -1,8 +1,12 @@
-"""Initialize common definitions for kinematics."""
+"""Import common definitions for kinematics."""
 
-from typing import Dict
-
-DEFAULT_FRAME = "map"
-
-Configuration = Dict[str, float]
-"""A map from joint names to positions (rad or m)."""
+from .kinematics_core import DEFAULT_FRAME as DEFAULT_FRAME
+from .kinematics_core import Configuration as Configuration
+from .point3d import Point3D as Point3D
+from .poses import Pose2D as Pose2D
+from .poses import Pose3D as Pose3D
+from .rotations import EulerRPY as EulerRPY
+from .rotations import Quaternion as Quaternion
+from .trajectories import CartesianPath as CartesianPath
+from .trajectories import Path as Path
+from .waypoints import Waypoints as Waypoints
