+++ src/robotics_utils/kinematics/kinematic_tree.py
@@ -5,11 +5,9 @@ from __future__ import annotations
 from pathlib import Path
 from typing import Any
 
-from robotics_utils.filesystem.yaml_utils import load_named_poses, load_yaml_data
-from robotics_utils.kinematics import DEFAULT_FRAME, Configuration
-from robotics_utils.kinematics.collision_models import CollisionModel
-from robotics_utils.kinematics.poses import Pose3D
-from robotics_utils.kinematics.waypoints import Waypoints
+from robotics_utils.collision_models import CollisionModel
+from robotics_utils.io.yaml_utils import load_yaml_data
+from robotics_utils.kinematics import DEFAULT_FRAME, Configuration, Pose3D, Waypoints
 
 
 class KinematicTree:
@@ -49,10 +47,10 @@ class KinematicTree:
 
         tree = KinematicTree(root_frame=default_frame)
 
-        for obj_name, obj_pose in load_named_poses(yaml_path, "object_poses").items():
+        for obj_name, obj_pose in Pose3D.load_named_poses(yaml_path, "object_poses").items():
             tree.set_object_pose(obj_name, obj_pose)
 
-        for robot_name, base_pose in load_named_poses(yaml_path, "robot_base_poses").items():
+        for robot_name, base_pose in Pose3D.load_named_poses(yaml_path, "robot_base_poses").items():
             tree.set_robot_base_pose(robot_name, base_pose)
             tree.robot_configurations[robot_name] = {}  # Default: No robot configurations in YAML
