+++ src/robotics_utils/math/conversions.py
@@ -1,0 +1,4 @@
+"""Define constants for unit conversion."""
+
+Gi_TO_B = 1024**3
+"""Convert from gibibytes (GiB) to bytes (B)."""
