+++ src/robotics_utils/ros/transform_recorder.py
@@ -6,7 +6,7 @@ from pathlib import Path
 
 import rospy
 
-from robotics_utils.filesystem.yaml_utils import export_yaml_data
+from robotics_utils.io.yaml_utils import export_yaml_data
 from robotics_utils.kinematics.poses import Pose3D
 from robotics_utils.ros.transform_manager import TransformManager
