+++ src/robotics_utils/sensors/realsense.py
@@ -12,7 +12,6 @@ from rich.console import Console
 from rich.table import Table
 from typing_extensions import Self
 
-from robotics_utils.filesystem.logging import log_info
 from robotics_utils.sensors.cameras import CameraIntrinsics, DepthCameraSpec, Resolution
 from robotics_utils.vision.images import DepthImage, RGBDImage, RGBImage
